Jd/adjacency matrix

+++ src/ptdf_calculations.jl
@@ -13,7 +13,7 @@ end
 function _buildptdf(
     branches,
     nodes::Vector{PSY.Bus},
-    bus_lookup::Dict{Int64, Int64},
+    bus_lookup::Dict{Int, Int},
     dist_slack::Vector{Float64},
     linear_solver::String = "Dense")
     if linear_solver == "KLU"
@@ -29,7 +29,7 @@ end
 
 function _buildptdf_from_matrices(
     A::IncidenceMatrix,
-    BA::SparseArrays.SparseMatrixCSC{T, Int32} where {T <: Union{Float32, Float64}},
+    BA::SparseArrays.SparseMatrixCSC{T, Int} where {T <: Union{Float32, Float64}},
     dist_slack::Vector{Float64},
     linear_solver::String)
     if linear_solver == "KLU"
@@ -51,9 +51,9 @@ end
 
 # PTDF evaluation ############################################################
 function _calculate_PTDF_matrix_KLU(
-    A::SparseArrays.SparseMatrixCSC{Int8, Int32},
-    BA::SparseArrays.SparseMatrixCSC{Float64, Int32},
-    slack_positions::Vector{Int64},
+    A::SparseArrays.SparseMatrixCSC{Int8, Int},
+    BA::SparseArrays.SparseMatrixCSC{Float64, Int},
+    slack_positions::Vector{Int},
     dist_slack::Vector{Float64})
     linecount = size(BA, 1)
     buscount = size(BA, 2)
@@ -83,7 +83,7 @@ end
 function calculate_PTDF_matrix_KLU(
     branches,
     nodes::Vector{PSY.Bus},
-    bus_lookup::Dict{Int64, Int64},
+    bus_lookup::Dict{Int, Int},
     dist_slack::Vector{Float64})
     A, slack_positions = calculate_A_matrix(branches, nodes)
     BA = calculate_BA_matrix(branches, slack_positions, bus_lookup)
@@ -107,7 +107,7 @@ end
 function _calculate_PTDF_matrix_DENSE(
     A::Matrix{Int8},
     BA::Matrix{T},
-    slack_position::Vector{Int64},
+    slack_position::Vector{Int},
     dist_slack::Vector{Float64}) where {T <: Union{Float32, Float64}}
 
     # Use dense calculation of ABA
@@ -160,7 +160,7 @@ end
 function calculate_PTDF_matrix_DENSE(
     branches,
     nodes::Vector{PSY.Bus},
-    bus_lookup::Dict{Int64, Int64},
+    bus_lookup::Dict{Int, Int},
     dist_slack::Vector{Float64})
     A, slack_positions = calculate_A_matrix(branches, nodes)
     BA = Matrix(calculate_BA_matrix(branches, slack_positions, bus_lookup))
@@ -169,9 +169,9 @@ function calculate_PTDF_matrix_DENSE(
 end
 
 function _calculate_PTDF_matrix_MKLPardiso(
-    A::SparseArrays.SparseMatrixCSC{Int8, Int32},
-    BA::SparseArrays.SparseMatrixCSC{Float64, Int32},
-    slack_positions::Vector{Int64},
+    A::SparseArrays.SparseMatrixCSC{Int8, Int},
+    BA::SparseArrays.SparseMatrixCSC{Float64, Int},
+    slack_positions::Vector{Int},
     dist_slack::Vector{Float64})
     ps = Pardiso.MKLPardisoSolver()
 
@@ -202,7 +202,7 @@ end
 function calculate_PTDF_matrix_MKLPardiso(
     branches,
     nodes::Vector{PSY.Bus},
-    bus_lookup::Dict{Int64, Int64},
+    bus_lookup::Dict{Int, Int},
     dist_slack::Vector{Float64})
     A, slack_positions = calculate_A_matrix(branches, nodes)
     BA = calculate_BA_matrix(branches, slack_positions, bus_lookup)
